requirements.txt

matplotlib
matplotlib-inline

mrmr-selection

numpy

pandas
pandas-flavor

PyYAML

scikit-image
scikit-learn
scipy
seaborn

setuptools

torch
skorch

xgboost

